wip: update authentication system for team development
