--- g_league_dbt/streamlit_app/test_min.py
@@ -1,3 +1,0 @@
-import streamlit as st
-
-st.write("Hello World!")
