+++ requirements.txt
@@ -84,7 +84,6 @@ python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version <= "3
 python-slugify==8.0.4 ; python_version >= "3.10" and python_version <= "3.11" or python_version >= "3.12" and python_version < "3.13"
 pytimeparse==1.1.8 ; python_version >= "3.10" and python_version <= "3.11" or python_version >= "3.12" and python_version < "3.13"
 pytz==2025.2 ; python_version >= "3.10" and python_version <= "3.11" or python_version >= "3.12" and python_version < "3.13"
-pywin32==311 ; python_version >= "3.10" and python_version <= "3.11" and sys_platform == "win32" and platform_python_implementation != "PyPy" or python_version >= "3.12" and python_version < "3.13" and sys_platform == "win32" and platform_python_implementation != "PyPy"
 pyyaml==6.0.2 ; python_version >= "3.10" and python_version <= "3.11" or python_version >= "3.12" and python_version < "3.13"
 referencing==0.36.2 ; python_version >= "3.10" and python_version <= "3.11" or python_version >= "3.12" and python_version < "3.13"
 requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version <= "3.11" or python_version >= "3.12" and python_version < "3.13"
